backend/pyproject.toml
[project]
name = "arxiv-radar"
version = "0.1.0"
description = "Modern arXiv paper discovery with semantic search and personalized recommendations"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pgvector>=0.3",
    "sentence-transformers>=3.3",
    "redis>=5.2",
    "bcrypt>=4.2",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.18",
    "arxiv>=2.1",
    "apscheduler>=3.10",
    "httpx>=0.28",
    "email-validator>=2.2",
]

[project.urls]
Repository = "https://github.com/yourname/arxiv-radar"
Issues = "https://github.com/yourname/arxiv-radar/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
